common base class
